+++ crates/cxx-qt-gen/test_outputs/properties.cpp
@@ -5,7 +5,7 @@
 namespace rust::cxxqt1 {
 template<>
 SignalHandler<
-  ::rust::cxxqtgen1::cxx_qt::my_object::
+  ::cxx_qt::my_object::rust::cxxqtgen1::
     MyObjectCxxQtSignalParamsprimitiveChanged*>::~SignalHandler() noexcept
 {
   if (data[0] == nullptr && data[1] == nullptr) {
@@ -18,52 +18,52 @@ SignalHandler<
 template<>
 template<>
 void
-SignalHandler<::rust::cxxqtgen1::cxx_qt::my_object::
+SignalHandler<::cxx_qt::my_object::rust::cxxqtgen1::
                 MyObjectCxxQtSignalParamsprimitiveChanged*>::
-operator()<::cxx_qt::my_object::MyObject&>(::cxx_qt::my_object::MyObject& self)
+operator()<cxx_qt::my_object::MyObject&>(cxx_qt::my_object::MyObject& self)
 {
   call_MyObject_signal_handler_primitiveChanged(*this, self);
 }
 
 static_assert(
-  alignof(SignalHandler<::rust::cxxqtgen1::cxx_qt::my_object::
+  alignof(SignalHandler<::cxx_qt::my_object::rust::cxxqtgen1::
                           MyObjectCxxQtSignalParamsprimitiveChanged*>) <=
     alignof(::std::size_t),
   "unexpected aligment");
 static_assert(
-  sizeof(SignalHandler<::rust::cxxqtgen1::cxx_qt::my_object::
+  sizeof(SignalHandler<::cxx_qt::my_object::rust::cxxqtgen1::
                          MyObjectCxxQtSignalParamsprimitiveChanged*>) ==
     sizeof(::std::size_t[2]),
   "unexpected size");
 } // namespace rust::cxxqt1
 
-namespace rust::cxxqtgen1::cxx_qt::my_object {
+namespace cxx_qt::my_object::rust::cxxqtgen1 {
 ::QMetaObject::Connection
 MyObject_primitiveChangedConnect(
-  ::cxx_qt::my_object::MyObject& self,
-  ::rust::cxxqtgen1::cxx_qt::my_object::
+  cxx_qt::my_object::MyObject& self,
+  ::cxx_qt::my_object::rust::cxxqtgen1::
     MyObjectCxxQtSignalHandlerprimitiveChanged closure,
   ::Qt::ConnectionType type)
 {
   return ::QObject::connect(
     &self,
-    &::cxx_qt::my_object::MyObject::primitiveChanged,
+    &cxx_qt::my_object::MyObject::primitiveChanged,
     &self,
     [&, closure = ::std::move(closure)]() mutable {
-      const ::rust::cxxqt1::MaybeLockGuard<::cxx_qt::my_object::MyObject> guard(
+      const ::rust::cxxqt1::MaybeLockGuard<cxx_qt::my_object::MyObject> guard(
         self);
-      closure.template operator()<::cxx_qt::my_object::MyObject&>(self);
+      closure.template operator()<cxx_qt::my_object::MyObject&>(self);
     },
     type);
 }
-} // namespace rust::cxxqtgen1::cxx_qt::my_object
+} // namespace cxx_qt::my_object::rust::cxxqtgen1
 
 // Define namespace otherwise we hit a GCC bug
 // https://gcc.gnu.org/bugzilla/show_bug.cgi?id=56480
 namespace rust::cxxqt1 {
 template<>
 SignalHandler<
-  ::rust::cxxqtgen1::cxx_qt::my_object::
+  ::cxx_qt::my_object::rust::cxxqtgen1::
     MyObjectCxxQtSignalParamstrivialChanged*>::~SignalHandler() noexcept
 {
   if (data[0] == nullptr && data[1] == nullptr) {
@@ -76,45 +76,45 @@ SignalHandler<
 template<>
 template<>
 void
-SignalHandler<::rust::cxxqtgen1::cxx_qt::my_object::
+SignalHandler<::cxx_qt::my_object::rust::cxxqtgen1::
                 MyObjectCxxQtSignalParamstrivialChanged*>::
-operator()<::cxx_qt::my_object::MyObject&>(::cxx_qt::my_object::MyObject& self)
+operator()<cxx_qt::my_object::MyObject&>(cxx_qt::my_object::MyObject& self)
 {
   call_MyObject_signal_handler_trivialChanged(*this, self);
 }
 
 static_assert(
-  alignof(SignalHandler<::rust::cxxqtgen1::cxx_qt::my_object::
+  alignof(SignalHandler<::cxx_qt::my_object::rust::cxxqtgen1::
                           MyObjectCxxQtSignalParamstrivialChanged*>) <=
     alignof(::std::size_t),
   "unexpected aligment");
 static_assert(
-  sizeof(SignalHandler<::rust::cxxqtgen1::cxx_qt::my_object::
+  sizeof(SignalHandler<::cxx_qt::my_object::rust::cxxqtgen1::
                          MyObjectCxxQtSignalParamstrivialChanged*>) ==
     sizeof(::std::size_t[2]),
   "unexpected size");
 } // namespace rust::cxxqt1
 
-namespace rust::cxxqtgen1::cxx_qt::my_object {
+namespace cxx_qt::my_object::rust::cxxqtgen1 {
 ::QMetaObject::Connection
 MyObject_trivialChangedConnect(
-  ::cxx_qt::my_object::MyObject& self,
-  ::rust::cxxqtgen1::cxx_qt::my_object::MyObjectCxxQtSignalHandlertrivialChanged
+  cxx_qt::my_object::MyObject& self,
+  ::cxx_qt::my_object::rust::cxxqtgen1::MyObjectCxxQtSignalHandlertrivialChanged
     closure,
   ::Qt::ConnectionType type)
 {
   return ::QObject::connect(
     &self,
-    &::cxx_qt::my_object::MyObject::trivialChanged,
+    &cxx_qt::my_object::MyObject::trivialChanged,
     &self,
     [&, closure = ::std::move(closure)]() mutable {
-      const ::rust::cxxqt1::MaybeLockGuard<::cxx_qt::my_object::MyObject> guard(
+      const ::rust::cxxqt1::MaybeLockGuard<cxx_qt::my_object::MyObject> guard(
         self);
-      closure.template operator()<::cxx_qt::my_object::MyObject&>(self);
+      closure.template operator()<cxx_qt::my_object::MyObject&>(self);
     },
     type);
 }
-} // namespace rust::cxxqtgen1::cxx_qt::my_object
+} // namespace cxx_qt::my_object::rust::cxxqtgen1
 
 namespace cxx_qt::my_object {
 ::std::int32_t const&