+++ crates/cxx-qt-lib/src/core/qbytearray.rs
@@ -112,9 +112,9 @@ pub struct QByteArray {
     ///
     /// Qt5 QByteArray has one pointer as a member
     /// Qt6 QByteArray has one member, which contains two pointers and a size_t
-    #[cfg(qt_version_major = "5")]
+    #[cfg(cxxqt_qt_version_major = "5")]
     _space: MaybeUninit<usize>,
-    #[cfg(qt_version_major = "6")]
+    #[cfg(cxxqt_qt_version_major = "6")]
     _space: MaybeUninit<[usize; 3]>,
 }